backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
apscheduler==3.10.4
openai==1.10.0
sqlalchemy==2.0.25
pymysql==1.1.0
marko==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
